kernel/wlan: Re-fix printf issues in debug messages.

+++ sys/netproto/802_11/ieee80211_input.h
@@ -45,8 +45,8 @@
 #define	IEEE80211_VERIFY_LENGTH(_len, _minlen, _action) do {		\
 	if ((_len) < (_minlen)) {					\
 		IEEE80211_DISCARD(vap, IEEE80211_MSG_ELEMID,		\
-		    wh, NULL, "ie too short, got %d, expected %d",	\
-		    (_len), (_minlen));					\
+		    wh, NULL, "ie too short, got %ld, expected %ld",	\
+		    (long)(_len), (long)(_minlen));			\
 		vap->iv_stats.is_rx_elem_toosmall++;			\
 		_action;						\
 	}								\
